Consolidate CLI package by moving API and VM directories under it
