--- api/index.ts
@@ -1,25 +1,0 @@
-import { build, Builder } from "df/api/commands/build";
-import { compile } from "df/api/commands/compile";
-import * as credentials from "df/api/commands/credentials";
-import { init } from "df/api/commands/init";
-import { install } from "df/api/commands/install";
-import { prune } from "df/api/commands/prune";
-import * as query from "df/api/commands/query";
-import { run, Runner } from "df/api/commands/run";
-import * as table from "df/api/commands/table";
-import { test } from "df/api/commands/test";
-
-export {
-  init,
-  install,
-  credentials,
-  compile,
-  test,
-  build,
-  run,
-  query,
-  table,
-  Runner,
-  Builder,
-  prune
-};
