+++ cli/api/BUILD
@@ -9,7 +9,7 @@ ts_library(
         exclude = ["utils/**/*.*"],
     ),
     deps = [
-        "//api/utils",
+        "//cli/api/utils",
         "//common/errors",
         "//common/flags",
         "//common/promises",
@@ -17,9 +17,9 @@ ts_library(
         "//common/strings",
         "//core",
         "//protos:ts",
-        "//sandbox/vm:compile_loader",
+        "//cli/vm:compile_loader",
         "//sqlx",
-        "//sandbox/vm:vm",
+        "//cli/vm:vm",
         "@npm//@google-cloud/bigquery",
         "@npm//@types/glob",
         "@npm//@types/js-beautify",
