+++ cli/api/commands/install.ts
