+++ cli/api/commands/prune.ts
