+++ cli/api/commands/tslint.json
