+++ cli/api/index.ts
@@ -1,0 +1,25 @@
+import { build, Builder } from "df/cli/api/commands/build";
+import { compile } from "df/cli/api/commands/compile";
+import * as credentials from "df/cli/api/commands/credentials";
+import { init } from "df/cli/api/commands/init";
+import { install } from "df/cli/api/commands/install";
+import { prune } from "df/cli/api/commands/prune";
+import * as query from "df/cli/api/commands/query";
+import { run, Runner } from "df/cli/api/commands/run";
+import * as table from "df/cli/api/commands/table";
+import { test } from "df/cli/api/commands/test";
+
+export {
+  init,
+  install,
+  credentials,
+  compile,
+  test,
+  build,
+  run,
+  query,
+  table,
+  Runner,
+  Builder,
+  prune
+};
