+++ cli/vm/BUILD
