+++ tests/integration/bigquery.spec.ts
@@ -1,10 +1,10 @@
 import { expect } from "chai";
 import Long from "long";
 
-import * as dfapi from "df/api";
-import * as dbadapters from "df/api/dbadapters";
-import { BigQueryDbAdapter } from "df/api/dbadapters/bigquery";
-import { ExecutionSql } from "df/api/dbadapters/execution_sql";
+import * as dfapi from "df/cli/api";
+import * as dbadapters from "df/cli/api/dbadapters";
+import { BigQueryDbAdapter } from "df/cli/api/dbadapters/bigquery";
+import { ExecutionSql } from "df/cli/api/dbadapters/execution_sql";
 import { targetAsReadableString } from "df/core/targets";
 import { dataform } from "df/protos/ts";
 import { suite, test } from "df/testing";
